Refactor HomePage layout by removing FeaturesSection and integrating FeatureCard into SolutionSection
